Allow starting and stopping, use mtl (#39)
* Allow starting and stopping tasks

* Use mtl (and don't use RIO)

* Automatically remove finished tasks and events

* Reduce repetition in Plan.hs

* Refactor Functions.hs

* Make task planning include day of deadline

* Print error message if no tasks planned

Not if no tasks are defined, as before. This means that if there are
future events, the message is still printed. Also, the message for
invalid commands comes before it, and it still allows tasks to be
removed.

* Change `plan rm` message

No mentioning "task", because it can be an event
"Removed" -> "Removing" to fit with "Adding"

* Remove commented out code